fix(web-app): unread notifications are older than the default time range
