+++ packages/cube-frontend-web-app/src/hooks/usePollNotifications/useHasUnreadNotifications.ts
@@ -2,7 +2,11 @@ import { NotificationsApiGetLastNotificationRequest } from '@cube-frontend/api'
 import { notificationsApi } from '@cube-frontend/web-app/api/cosApi'
 import { DataCenterContext } from '@cube-frontend/web-app/context/DataCenterContext'
 import { NotificationsContext } from '@cube-frontend/web-app/context/NotificationsContext'
-import { checkIsNotificationUnread } from '@cube-frontend/web-app/utils/notification'
+import {
+  checkIsNotificationUnread,
+  checkIsNotificationWithin,
+  ListNotificationsPastEnum,
+} from '@cube-frontend/web-app/utils/notification'
 import { useContext } from 'react'
 import { useCosGetRequest } from '../useCosRequest/useCosGetRequest'
 import { usePolling } from '../usePolling'
@@ -27,6 +31,16 @@ export const useHasUnreadNotifications = (): boolean => {
 
   return (
     !!lastNotificationTime &&
+    /**
+     * The widest time range in the dropdown is 30 days, so notifications older
+     * than this should not appear in the table, and the unread indicator in the
+     * function bar should should also be hidden if the last notification was
+     * created more than 30 days ago.
+     */
+    checkIsNotificationWithin(
+      lastNotificationTime,
+      ListNotificationsPastEnum['30d'],
+    ) &&
     checkIsNotificationUnread(lastNotificationTime, lastAccessedAt)
   )
 }
